refactor for specialize noTx save implementation for embedded and remote operations
