+++ core/src/main/java/com/orientechnologies/orient/core/db/ODatabaseDocumentInternal.java
@@ -115,14 +115,8 @@ public interface ODatabaseDocumentInternal extends ODatabaseSession, ODatabaseIn
       final boolean ignoreCache, final boolean iUpdateCache, final boolean loadTombstones,
       final OStorage.LOCKING_STRATEGY lockingStrategy, RecordReader recordReader);
 
-  <RET extends ORecord> RET executeSaveRecord(final ORecord record, String clusterName, final int ver, final OPERATION_MODE mode,
-      boolean forceCreate, final ORecordCallback<? extends Number> recordCreatedCallback,
-      ORecordCallback<Integer> recordUpdatedCallback);
-
   void executeDeleteRecord(OIdentifiable record, final int iVersion, final boolean iRequired, final OPERATION_MODE iMode,
       boolean prohibitTombstones);
-
-  <RET extends ORecord> RET executeSaveEmptyRecord(ORecord record, String clusterName);
 
   void setDefaultTransactionMode();
 
@@ -239,5 +233,10 @@ public interface ODatabaseDocumentInternal extends ODatabaseSession, ODatabaseIn
   boolean isClusterVertex(int cluster);
 
   boolean isClusterEdge(int cluster);
+
+  ORecord saveAll(ORecord iRecord, String iClusterName, OPERATION_MODE iMode, boolean iForceCreate,
+      ORecordCallback<? extends Number> iRecordCreatedCallback, ORecordCallback<Integer> iRecordUpdatedCallback);
+
+  String getClusterName(final ORecord record);
 
 }
